requirements.txt
# Python==3.8.7

bcrypt==4.0.1
cryptography==41.0.2
mysql-connector-python==8.0.33
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
